# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/job": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取定时任务列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "定时任务管理"
                ],
                "summary": "获取定时任务列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "是否需要分页， 默认需要， 如果不分页 传 true",
                        "name": "isNotPaging",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "页码， 如果不分页 传 0",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "一页大小， 如果不分页 传 0",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "任务名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "任务描述",
                        "name": "mark",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code: 200 成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/res.Full"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.ResQueryJob"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "错误返回内容",
                        "schema": {
                            "$ref": "#/definitions/res.Base"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "修改定时任务数据",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "定时任务管理"
                ],
                "summary": "修改定时任务数据",
                "parameters": [
                    {
                        "description": "修改定时任务数据",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ReqModifyJob"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code: 200 成功",
                        "schema": {
                            "$ref": "#/definitions/res.Base"
                        }
                    },
                    "500": {
                        "description": "错误返回内容",
                        "schema": {
                            "$ref": "#/definitions/res.Base"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "添加定时任务数据",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "定时任务管理"
                ],
                "summary": "添加定时任务数据",
                "parameters": [
                    {
                        "description": "添加定时任务数据",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ReqCreateJob"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code: 200 成功",
                        "schema": {
                            "$ref": "#/definitions/res.Base"
                        }
                    },
                    "500": {
                        "description": "错误返回内容",
                        "schema": {
                            "$ref": "#/definitions/res.Base"
                        }
                    }
                }
            }
        },
        "/api/job/status/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "修改定时任务数据",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "定时任务管理"
                ],
                "summary": "修改定时任务数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code: 200 成功",
                        "schema": {
                            "$ref": "#/definitions/res.Base"
                        }
                    },
                    "500": {
                        "description": "错误返回内容",
                        "schema": {
                            "$ref": "#/definitions/res.Base"
                        }
                    }
                }
            }
        },
        "/api/job/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除定时任务",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "定时任务管理"
                ],
                "summary": "删除定时任务",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code: 200 成功",
                        "schema": {
                            "$ref": "#/definitions/res.Base"
                        }
                    },
                    "500": {
                        "description": "错误返回内容",
                        "schema": {
                            "$ref": "#/definitions/res.Base"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.ReqCreateJob": {
            "type": "object",
            "required": [
                "area_id",
                "expr",
                "name",
                "path"
            ],
            "properties": {
                "area_id": {
                    "description": "参数域ID",
                    "type": "integer",
                    "example": 10000
                },
                "expr": {
                    "description": "时间表达式",
                    "type": "string",
                    "example": "0/5 * * * * ?"
                },
                "mark": {
                    "description": "备注",
                    "type": "string",
                    "example": "每五秒执行一次脚本"
                },
                "name": {
                    "description": "名称",
                    "type": "string",
                    "example": "测试定时任务"
                },
                "path": {
                    "description": "脚本路径",
                    "type": "string",
                    "example": "test.js"
                }
            }
        },
...
        "model.ReqModifyJob": {
            "type": "object",
            "required": [
                "area_id",
                "expr",
                "id",
                "name",
                "path"
            ],
            "properties": {
                "area_id": {
                    "description": "参数域ID",
                    "type": "integer",
                    "example": 10000
                },
                "expr": {
                    "description": "时间表达式",
                    "type": "string"
                },
                "id": {
                    "description": "任务ID",
                    "type": "integer"
                },
                "mark": {
                    "description": "备注",
                    "type": "string"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "path": {
                    "description": "脚本路径",
                    "type": "string"
                }
            }
        },
...
        "model.ResQueryJob": {
            "type": "object",
            "properties": {
                "area": {
                    "description": "参数域",
                    "type": "string"
                },
                "area_id": {
                    "type": "integer"
                },
                "create_time": {
                    "description": "创建时间",
                    "type": "string"
                },
                "enable": {
                    "description": "状态",
                    "type": "boolean"
                },
                "expr": {
                    "description": "时间表达式",
                    "type": "string"
                },
                "id": {
                    "description": "任务ID",
                    "type": "integer"
                },
                "mark": {
                    "description": "备注",
                    "type": "string"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "path": {
                    "description": "脚本路径",
                    "type": "string"
                }
            }
        },
        "res.Base": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "allOf": [
                        {
                            "$ref": "#/definitions/res.RCode"
                        }
                    ]
                },
                "message": {
                    "description": "状态消息",
                    "type": "string"
                },
                "requestDatetime": {
                    "description": "请求时间",
                    "type": "string"
                },
                "requestId": {
                    "description": "请求ID",
                    "type": "string"
                },
                "responseDatetime": {
                    "description": "返回时间",
                    "type": "string"
                }
            }
        },
        "res.Full": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "allOf": [
                        {
                            "$ref": "#/definitions/res.RCode"
                        }
                    ]
                },
                "data": {
                    "description": "返回数据"
                },
                "message": {
                    "description": "状态消息",
                    "type": "string"
                },
                "requestDatetime": {
                    "description": "请求时间",
                    "type": "string"
                },
                "requestId": {
                    "description": "请求ID",
                    "type": "string"
                },
                "responseDatetime": {
                    "description": "返回时间",
                    "type": "string"
                }
            }
        },
        "res.RCode": {
            "type": "integer",
            "enum": [
                200,
                201,
                400,
                401,
                404,
                405,
                500,
                501,
                502,
                504,
                505
            ],
            "x-enum-varnames": [
                "OK",
                "OK_CREATE",
                "BAD_REQUEST",
                "UNAUTHORIZED",
                "NOT_FOUND",
                "METHOD_NOT_ALLOWED",
                "INTERNAL_SERVER_ERROR",
                "NOT_IMPLEMENTED",
                "BAD_GATEWAY",
                "GATEWAY_TIMEOUT",
                "HTTP_VERSION_NOT_SUPPORTED"
            ]
        }
    }
}